Move DB out of internal for use in tools that need to read the content
