+++ clone/logdb/database_test.go
